Move trace category settings to timeline events tab bar
