+++ packages/devtools_app/lib/src/screens/performance/tabbed_performance_view.dart
@@ -8,13 +8,10 @@ import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 
 import '../../shared/analytics/constants.dart' as gac;
-import '../../shared/charts/flame_chart.dart';
 import '../../shared/common_widgets.dart';
 import '../../shared/feature_flags.dart';
 import '../../shared/globals.dart';
 import '../../shared/primitives/auto_dispose.dart';
-import '../../shared/theme.dart';
-import '../../shared/ui/search.dart';
 import '../../shared/ui/tab.dart';
 import '../../shared/utils.dart';
 import 'panes/flutter_frames/flutter_frame_model.dart';
@@ -22,12 +19,8 @@ import 'panes/flutter_frames/flutter_frames_controller.dart';
 import 'panes/frame_analysis/frame_analysis.dart';
 import 'panes/raster_stats/raster_stats.dart';
 import 'panes/rebuild_stats/rebuild_stats.dart';
-import 'panes/timeline_events/legacy/legacy_events_controller.dart';
-import 'panes/timeline_events/legacy/timeline_flame_chart.dart';
-import 'panes/timeline_events/perfetto/perfetto.dart';
-import 'panes/timeline_events/timeline_events_controller.dart';
+import 'panes/timeline_events/timeline_events_view.dart';
 import 'performance_controller.dart';
-import 'performance_screen.dart';
 
 class TabbedPerformanceView extends StatefulWidget {
   const TabbedPerformanceView();
@@ -44,8 +37,6 @@ class _TabbedPerformanceViewState extends State<TabbedPerformanceView>
 
   late FlutterFramesController _flutterFramesController;
 
-  late TimelineEventsController _timelineEventsController;
-
   FlutterFrame? _selectedFlutterFrame;
 
   @override
@@ -53,7 +44,6 @@ class _TabbedPerformanceViewState extends State<TabbedPerformanceView>
     super.didChangeDependencies();
     if (!initController()) return;
 
-    _timelineEventsController = controller.timelineEventsController;
     _flutterFramesController = controller.flutterFramesController;
 
     cancelListeners();
@@ -184,68 +174,12 @@ class _TabbedPerformanceViewState extends State<TabbedPerformanceView>
     return _PerformanceTabRecord(
       tab: _buildTab(
         tabName: 'Timeline Events',
-        trailing: ValueListenableBuilder<bool>(
-          valueListenable: _timelineEventsController.useLegacyTraceViewer,
-          builder: (context, useLegacy, _) {
-            return Row(
-              mainAxisAlignment: MainAxisAlignment.end,
-              children: [
-                if (useLegacy || !FeatureFlags.embeddedPerfetto) ...[
-                  ValueListenableBuilder<EventsControllerStatus>(
-                    valueListenable: _timelineEventsController.status,
-                    builder: (context, status, _) {
-                      final searchFieldEnabled =
-                          status == EventsControllerStatus.ready;
-                      return SearchField<LegacyTimelineEventsController>(
-                        searchController:
-                            _timelineEventsController.legacyController,
-                        searchFieldEnabled: searchFieldEnabled,
-                        searchFieldWidth: wideSearchFieldWidth,
-                      );
-                    },
-                  ),
-                  const SizedBox(width: denseSpacing),
-                  FlameChartHelpButton(
-                    gaScreen: PerformanceScreen.id,
-                    gaSelection: gac.timelineFlameChartHelp,
-                  ),
-                ],
-                if (!offlineController.offlineMode.value)
-                  const SizedBox(width: denseSpacing),
-                RefreshTimelineEventsButton(
-                  controller: _timelineEventsController,
-                ),
-              ],
-            );
-          },
+        trailing: TimelineEventsTabControls(
+          controller: controller.timelineEventsController,
         ),
       ),
-      tabView: ValueListenableBuilder<bool>(
-        valueListenable: _timelineEventsController.useLegacyTraceViewer,
-        builder: (context, useLegacy, _) {
-          return (useLegacy || !FeatureFlags.embeddedPerfetto)
-              ? KeepAliveWrapper(
-                  child: DualValueListenableBuilder<EventsControllerStatus,
-                      double>(
-                    firstListenable: _timelineEventsController.status,
-                    secondListenable: _timelineEventsController
-                        .legacyController.processor.progressNotifier,
-                    builder: (context, status, processingProgress, _) {
-                      return TimelineEventsView(
-                        controller: _timelineEventsController,
-                        processing: status == EventsControllerStatus.processing,
-                        processingProgress: processingProgress,
-                      );
-                    },
-                  ),
-                )
-              : KeepAliveWrapper(
-                  child: EmbeddedPerfetto(
-                    perfettoController:
-                        _timelineEventsController.perfettoController,
-                  ),
-                );
-        },
+      tabView: TimelineEventsTabView(
+        controller: controller.timelineEventsController,
       ),
       featureController: controller.timelineEventsController,
     );
@@ -256,35 +190,6 @@ class _TabbedPerformanceViewState extends State<TabbedPerformanceView>
       tabName: tabName,
       gaPrefix: _gaPrefix,
       trailing: trailing,
-    );
-  }
-}
-
-@visibleForTesting
-class RefreshTimelineEventsButton extends StatelessWidget {
-  const RefreshTimelineEventsButton({
-    Key? key,
-    required this.controller,
-  }) : super(key: key);
-
-  final TimelineEventsController controller;
-
-  @override
-  Widget build(BuildContext context) {
-    return ValueListenableBuilder<EventsControllerStatus>(
-      valueListenable: controller.status,
-      builder: (context, status, _) {
-        return RefreshButton(
-          iconOnly: true,
-          outlined: false,
-          onPressed: status == EventsControllerStatus.processing
-              ? null
-              : controller.processAllTraceEvents,
-          tooltip: 'Refresh timeline events',
-          gaScreen: gac.performance,
-          gaSelection: gac.refreshTimelineEvents,
-        );
-      },
     );
   }
 }
